backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.2
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
langchain-openai==0.2.14
langgraph==0.2.43
certifi==2025.1.31
